server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "GraphQL server for events generated by Apibara indexer"
authors = []

[tool.poetry.scripts]
server = "server.cli:run"

[tool.poetry.dependencies]
python = "3.11.3"
strawberry-graphql = {extras = ["debug-server"], version = "^0.216.1"}
aiohttp-cors = "^0.7.0"
aiohttp = "^3.9.1"
pymongo = "^4.6.1"
starknet-py = "^0.18.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
